requirements.txt
Flask==3.0.0
Werkzeug>=3.0.0
requests==2.31.0
pytest==8.3.3
SpeechRecognition==3.14.3
PyAudio==0.2.14
libsass==0.23.0
python-dotenv==1.0.0
openai>=1.50.0
httpx==0.27.2
sentry-sdk[flask]==1.40.0
sqlalchemy==2.0.25
sqlalchemy-utils==0.41.1
google-generativeai==0.8.3

